fix: send zero-value arguments if calculating subsidy of genesis block
Dash uses the height and difficulty of the previous block to calculate
the subsidy for the current block... which in the case of the genesis
block is block -1, which doesn't exist.

Attempting in reading `pprev` which is will evaluate to a `nullptr`, so
for any blocks <=0, we have to use precalculated value.