chore: wallet_upgadewallet.py to use the oldest supported dash version
Version is changed from 19.3.0 to 0.16.1.1
Even older doesn't support importing private coinbase key at the moment.

Version 20.x is also changed to 18.2 to the latest which do not create wallets
by default for sack of unification with bitcoin and to fix issue with `createwallet`
rpc which is subject to change once descriptor wallets would be backported.