build: don't error on `-Wattributes` with GCC <14
We need to annotate `CHDPubKey::nVersion` with `[[maybe_unused]]` to
avoid upsetting Clang due to `-Wunused-private-field`. GCC 11 doesn't
seem to emit a warning for that variable and therefore, finds the
annotation unnecessary, triggering an error due to `-Wattributes`.

Since both Clang 18 and GCC 14 agree with the `-Wunused-private-field`
assessment, we are better off suppressing `-Wattributes` error-like
treatment on lower compiler versions.